+++ assets/back-end/js/shortcode-modal.js
@@ -14,6 +14,7 @@ var CVMVideo_DIALOG_WIN = false;
 			}
 			
 			var theme = $('#cvm_playlist_theme').val(),
+				order = $('#cvm_playlist_posts_order').val(),
 				aspect = $('#aspect_ratio').val(),
 				width = $('#width').val(),
 				volume = $('#volume').val(),
@@ -25,6 +26,7 @@ var CVMVideo_DIALOG_WIN = false;
 			
 			if( 'default' == theme ){
 				extra_theme_opts += 'layout="' + $('input[name=cvm_theme_default_layout]:checked').val() + '"';
+				extra_theme_opts += ' show_excerpts="' + ( $('input[name=cvm_theme_default_show_excerpts]').is(':checked') ? 1 : 0 ) + '"';
 			}
 
 			var videos_array = $.grep( videos.split('|'), function(val){ return '' != val }),
@@ -48,7 +50,7 @@ var CVMVideo_DIALOG_WIN = false;
 				}
 			}
 
-			var	shortcode 	= '[cvm_playlist theme="' + theme + '" '+ extra_theme_opts +' aspect_ratio="' + aspect + '" width="' + width + '" volume="' + volume + '" title="'+title+'" byline="'+byline+'" portrait="'+portrait+'" playlist_loop="' + playlist_loop + '"' + s_videos + s_categories + s_post_types + ']';;
+			var	shortcode 	= '[cvm_playlist theme="' + theme + '" '+ extra_theme_opts + ' order="' + order + '"' + ' aspect_ratio="' + aspect + '" width="' + width + '" volume="' + volume + '" title="'+title+'" byline="'+byline+'" portrait="'+portrait+'" playlist_loop="' + playlist_loop + '"' + s_videos + s_categories + s_post_types + ']';;
 			
 			$('#cvm-playlist-items').find('input[name=cvm_selected_items]').val('');
 			$('#cvm-list-items').empty().html(CVM_SHORTCODE_MODAL.no_videos);