+++ includes/libs/admin/editor/classic-editor.class.php
@@ -355,21 +355,48 @@ class Classic_Editor{
                                         <label for=""><input type="radio" name="cvm_theme_default_layout" value="" checked="checked" /> <?php _e( 'bottom', 'codeflavors-vimeo-video-post-lite' );?></label>
                                         <label for=""><input type="radio" name="cvm_theme_default_layout" value="right" /> <?php _e( 'right', 'codeflavors-vimeo-video-post-lite' );?></label>
                                         <label for=""><input type="radio" name="cvm_theme_default_layout" value="left" /> <?php _e( 'left', 'codeflavors-vimeo-video-post-lite' );?></label>
+                                        <p>
+                                            <label>
+                                                <input type="checkbox" name="cvm_theme_default_show_excerpts" id="cvm_theme_default_show_excerpts" value="1" />
+                                                <?php _e('Show posts excerpts', 'codeflavors-vimeo-video-post-lite');?>
+                                            </label>
+                                        </p>
                                     </div>
 
 								</td>
 						</tr>
+                        <tr>
+                            <th><label for="cvm_posts_order"><?php _e( 'Posts order', 'codeflavors-vimeo-video-post-lite' );?>:</label></th>
+                            <td>
+	                            <?php
+	                            Helper_Admin::select(
+	                                [
+                                        'options' => [
+	                                        'manual' => __( 'Manual order', 'codeflavors-vimeo-video-post-lite' ),
+	                                        'newest' => __( 'Newest first', 'codeflavors-vimeo-video-post-lite' ),
+	                                        'oldest' => __( 'Oldest first', 'codeflavors-vimeo-video-post-lite' ),
+	                                        'alphabetical' => __( 'Alphabetically', 'codeflavors-vimeo-video-post-lite' )
+                                        ],
+                                        'name' => 'cvm_playlist_posts_order',
+                                        'id' => 'cvm_playlist_posts_order',
+                                        'selected' => 'manual',
+                                        'use_keys' => true
+	                                ]
+                                );
+	                            ?>
+                            </td>
+                        </tr>
 						<tr>
 							<th><label for="cvm_aspect_ratio"><?php _e('Aspect', 'codeflavors-vimeo-video-post-lite');?>:</label></th>
 							<td>
 								<?php
-		$args = [
-				'name' => 'aspect_ratio',
-				'id' => 'aspect_ratio',
-				'class' => 'cvm_aspect_ratio'
-		];
-		Helper_Admin::aspect_ratio_select( $args );
-		?>
+                                $args = [
+                                        'name' => 'aspect_ratio',
+                                        'id' => 'aspect_ratio',
+                                        'class' => 'cvm_aspect_ratio'
+                                ];
+                                Helper_Admin::aspect_ratio_select( $args );
+                                ?>
 								</td>
 						</tr>